Nightly build 4.1
Made a new class ‘initial_setup.java’ that adjusts the data based on
the center of mass. This is how the rest of the algorithm wants the
data to be set up, such that the center of mass becomes the origin. I
have realized that using LL data would through off calculation of new
nodes because it does not take into consideration actual distance but
instead relative distance. Although this can easily be adjusted later
once there is a proper mapping of the LL data to a 2d cartesian plane.