Nightly build 1200
Added new methods to List_ops.java for faster data structure conversion down the road. MST.java (renamed from MST_triplet.java) uses these new methods in List_ops.java to help the others take out data from the object, and put in new data to the object with whatever data structure they are more comfortable (be it an array or a List). New ‘Sum’ method added to Reduced_Kruskal.java, which is used to properly check if the first input is a logical adjacent matrix.